Integrated new single-pass selector evaluators, contributed by knz (Anton Kazennikov).
