+++ src/main/java/org/jsoup/select/Collector.java
@@ -1,7 +1,6 @@
 package org.jsoup.select;
 
 import org.jsoup.nodes.Element;
-import org.jsoup.nodes.Evaluator;
 import org.jsoup.nodes.Node;
 
 /**
@@ -22,15 +21,17 @@ public class Collector {
      */
     public static Elements collect (Evaluator eval, Element root) {
         Elements elements = new Elements();
-        new NodeTraversor(new Accumulator(elements, eval)).traverse(root);
+        new NodeTraversor(new Accumulator(root, elements, eval)).traverse(root);
         return elements;
     }
 
     private static class Accumulator implements NodeVisitor {
+        private final Element root;
         private final Elements elements;
         private final Evaluator eval;
 
-        Accumulator(Elements elements, Evaluator eval) {
+        Accumulator(Element root, Elements elements, Evaluator eval) {
+            this.root = root;
             this.elements = elements;
             this.eval = eval;
         }
@@ -38,7 +39,7 @@ public class Collector {
         public void head(Node node, int depth) {
             if (node instanceof Element) {
                 Element el = (Element) node;
-                if (eval.matches(el, el))
+                if (eval.matches(root, el))
                     elements.add(el);
             }
         }
