+++ src/main/java/org/jsoup/select/selectors/PreviousSiblingSelector.java
@@ -1,0 +1,40 @@
+package org.jsoup.select.selectors;
+
+import org.jsoup.nodes.Element;
+import org.jsoup.select.Evaluator;
+
+/**
+ * Tree-based evaluator for matching Element's parent.
+ * For evaluating 'E ~ F' construction
+ */
+public class PreviousSiblingSelector extends Evaluator {
+    Evaluator sel;
+
+    public PreviousSiblingSelector(Evaluator sel) {
+        this.sel = sel;
+    }
+
+    @Override
+    public boolean matches(Element root, Element element) {
+        if (root == element)
+            return false;
+
+        Element prev = element.previousElementSibling();
+
+        while (prev != null) {
+            if (sel.matches(root, prev))
+                return true;
+
+            prev = prev.previousElementSibling();
+        }
+
+        return false;
+    }
+
+    @Override
+    public String toString() {
+        return String.format(":prev*%s", sel);
+    }
+
+
+}
